move companion to the class its used in
